final commit

+++ src/modules/person/commands/handlers/delete-person.handler.ts
@@ -1,9 +1,9 @@
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
-import { SavePersonCommand } from '../impl/save-person.command';
-import { Person } from 'src/entities/person';
+import { Person } from '../../../../entities/person';
 import { DeletePersonCommand } from '../impl/delete-person.command';
+import { HttpException, HttpStatus } from '@nestjs/common';
 
 @CommandHandler(DeletePersonCommand)
 export class DeletePersonHandler
@@ -13,6 +13,15 @@ export class DeletePersonHandler
 
   async execute(command: DeletePersonCommand) {
     const id = command.id;
-    await this.repo.delete(id);
+    const person = await this.repo.findOne(id);
+    if (person) {
+      await this.repo.delete(id);
+      return { message: 'Person was deleted successfully!' };
+    } else {
+      throw new HttpException(
+        "Person with this id doesn't exist",
+        HttpStatus.NOT_FOUND,
+      );
+    }
   }
 }
